Add files via upload
Decorrelation is a general term for any process that is used to reduce autocorrelation within a signal, or cross-correlation within a set of signals, while preserving other aspects of the signal. In this project, we seek to demonstrate the process of decorrelation and examine its effects on the performance of predictive models.